ratings-data-loader/requirements.txt
google-cloud-bigquery==3.24.0